Add combination of theories
This is a simple implementation which runs propagate() to
exhastion (i.e., until no plugin propagates a new variable). This is
slightly wasteful with current implementation of propagate() in
plugins (bool, LRA). Plugins would ideally remember which
variables they have already processed so that they do not have
to check them again in subsequent calls.