Fix: do not reorder watched variables unnecessarily
In particular, do not swap watched variables, if the constraint is fully
assigned. We try to maintain an invariant that a constraint is not
fully assigned iff its first variable is unassigned. However, this could
be broken with the old code if we move variable assigned at a higher
decision level to the second position and then backtrack.