Fix: do not propagate equations eagerly
By eager propagation we mean, for example, setting x to 2 if there are
constraints `x == 2` or `2 <= x <= 2`. The original formulation of MCSat
expects that theory variables are set only by semantic *decisions*. There
are some complications with eagerly propagating implied equations.
However, it might be possible to do this and it could potentially be
beneficial.