Propagate literals retroactively in the bool plugin
If LRA propagates a constraint at a lower decision level, the bool plugin
should perform BCP to exhaustion including the retroactively propagated
literal. It is not an issue now because the only retroactively propagated
literals are new derivations which will not cause any further BCP. However,
it might be useful to retroactively propagate some old constraints in the
LRA plugin as well.