Trail triggers events in its resize() method
Variables are added by calling trail.resize(). However,
previously, it did not trigger the on_variable_resize event. The
event was called by the core solver when it discovered new
variables in a learned clause which is a bit confusing. Moreover,
new variables could only be added in a conflict.